requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
self==2020.12.3
contextlib2==0.6.0.post1
DateTime==4.3
discord==1.7.3
discord-components==2.1.2
discord.py==1.7.3
future==0.18.2
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
pefile==2021.5.24
Pillow==8.2.0
self==2020.12.3
pip==21.1.3
py==1.10.0
pyaml==20.4.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
schema==0.7.4
spline==1.12
typing-extensions==3.10.0.0
urllib3==1.26.5
utils==1.0.1
yarl==1.6.3
Cybernator==0.6.3
zope.interface==5.4.0
